requirements.txt
fastapi
uvicorn[standard]
requests
openpyxl
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
